refactor: Unify ios file system helpers
