+++ lib/commands/file-movement.js
@@ -1,10 +1,9 @@
 import _ from 'lodash';
-import { system, fs, tempDir, mkdirp, zip, util } from 'appium-support';
+import { fs, tempDir, mkdirp, zip, util } from 'appium-support';
 import path from 'path';
 import log from '../logger';
-import { exec } from 'teen_process';
-import B from 'bluebird';
 import { services } from 'appium-ios-device';
+import { pullFile, pullFolder, pushFile } from '../ios-fs-helpers';
 
 const CONTAINER_PATH_MARKER = '@';
 // https://regex101.com/r/PLdB0G/2
@@ -12,9 +11,7 @@ const CONTAINER_PATH_PATTERN = new RegExp(`^${CONTAINER_PATH_MARKER}([^/]+)/(.*)
 const CONTAINER_TYPE_SEPARATOR = ':';
 const IFUSE_CONTAINER_DOCUMENTS = 'documents';
 const CONTAINER_DOCUMENTS_PATH = 'Documents';
-const IO_TIMEOUT = 60000;
 const OBJECT_NOT_FOUND_ERROR_MESSAGE = 'OBJECT_NOT_FOUND';
-const MAX_PULL_CHUNK_SIZE = 5;
 
 const commands = {};
 
@@ -43,20 +40,6 @@ function isDocuments (containerType) {
   return _.toLower(containerType) === IFUSE_CONTAINER_DOCUMENTS;
 }
 
-async function mkdirpDevice (service, dir) {
-  if (dir === '.' || dir === '/') {
-    return;
-  }
-  try {
-    await service.listDirectory(dir);
-    return;
-  } catch (e) {
-    // This means that the directory is missing and we got an object not found error. Therefore, we are going to the parent
-    await mkdirpDevice(service, path.dirname(dir));
-  }
-  await service.createDirectory(dir);
-}
-
 async function createService (udid, remotePath) {
   if (CONTAINER_PATH_PATTERN.test(remotePath)) {
     const {bundleId, pathInContainer, containerType} = await parseContainerPath(remotePath);
@@ -65,76 +48,7 @@ async function createService (udid, remotePath) {
     return {service, relativePath};
   } else {
     const service = await createAfcClient(udid);
-    const relativePath = remotePath;
-    return {service, relativePath};
-  }
-}
-
-async function pullFileFromRealDevice (service, relativePath) {
-  const stream = await service.createReadStream(relativePath, { autoDestroy: true });
-  const pullPromise = new B((resolve, reject) => {
-    stream.on('close', resolve);
-    stream.on('error', reject);
-  });
-  const buffers = [];
-  stream.on('data', (data) => buffers.push(data));
-  try {
-    await pullPromise.timeout(IO_TIMEOUT);
-  } catch (e) {
-    throw new Error(`Couldn't pull the file '${relativePath}' ` +
-      `within the given timeout ${IO_TIMEOUT}ms. Original error: ${e.message}`);
-  }
-  return Buffer.concat(buffers).toString('base64');
-}
-
-async function pullFolderFromRealDevice (service, relativePath) {
-  const tmpFolder = await tempDir.openDir();
-  try {
-    const folderPath = path.join(tmpFolder, relativePath);
-    await mkdirp(folderPath);
-    const pullPromises = [];
-    const waitForPullChunks = async () => {
-      if (_.isEmpty(pullPromises)) {
-        return;
-      }
-
-      try {
-        await B.all(pullPromises).timeout(IO_TIMEOUT);
-      } catch (e) {
-        throw new Error(`Couldn't pull all items in the folder '${relativePath}' ` +
-          `within the given timeout ${IO_TIMEOUT}ms. Original error: ${e.message}`);
-      }
-    };
-    await service.walkDir(relativePath, true, async (itemPath, isDir) => {
-      const pathOnServer = path.join(tmpFolder, itemPath);
-      if (isDir) {
-        await fs.mkdir(pathOnServer);
-        return;
-      }
-
-      const readStream = await service.createReadStream(itemPath, {autoDestroy: true});
-      const writeStream = fs.createWriteStream(pathOnServer, {autoClose: true});
-      pullPromises.push(new B((resolve, reject) => {
-        writeStream.on('close', resolve);
-        const onStreamingError = (e) => {
-          readStream.unpipe(writeStream);
-          reject(e);
-        };
-        writeStream.on('error', onStreamingError);
-        readStream.on('error', onStreamingError);
-      }));
-      readStream.pipe(writeStream);
-      if (pullPromises.length % MAX_PULL_CHUNK_SIZE === 0) {
-        await waitForPullChunks();
-      }
-    });
-    // Wait for the rest of the chunks
-    await waitForPullChunks();
-    return (await zip.toInMemoryZip(folderPath, {
-      encodeToBase64: true,
-    })).toString();
-  } finally {
-    await fs.rimraf(tmpFolder);
+    return {service, relativePath: remotePath};
   }
 }
 
@@ -229,6 +143,30 @@ async function pushFileToSimulator (device, remotePath, base64Data) {
   }
 }
 
+// TODO: Add mobile extension to use this helper
+/**
+ * Get bundleIds which can mount by `--documents` flag
+ *
+ * @param {Object} udid - The udid of the target device
+ * @returns {Array<string>} A list of User level apps' bundle ids which has
+ *                          'UIFileSharingEnabled' attribute.
+ *                          Only user apps might have it.
+ */
+async function getAvailableBundleIds (udid) {
+  const service = await services.startInstallationProxyService(udid);
+  try {
+    const applications = await service.listApplications({applicationType: 'User'});
+    return _.toPairs(applications).reduce((acc, [bundleId, props]) => {
+      if (props.UIFileSharingEnabled) {
+        acc.push(bundleId);
+      }
+      return acc;
+    }, []);
+  } finally {
+    service.close();
+  }
+}
+
 /**
  * Save the given base64 data chunk as a binary file on the device under test.
  *
@@ -252,32 +190,19 @@ async function pushFileToSimulator (device, remotePath, base64Data) {
 async function pushFileToRealDevice (device, remotePath, base64Data) {
   const {service, relativePath} = await createService(device.udid, remotePath);
   try {
-    await mkdirpDevice(service, path.dirname(relativePath));
-    const stream = await service.createWriteStream(relativePath, {autoDestroy: true});
-    let pushError = null;
-    const pushPromise = new B((resolve, reject) => {
-      stream.on('error', (e) => {
-        pushError = e;
-      });
-      stream.on('close', () => {
-        if (pushError) {
-          reject(pushError);
-        } else {
-          resolve();
-        }
-      });
-    });
-    stream.write(Buffer.from(base64Data, 'base64'));
-    stream.end();
-    try {
-      await pushPromise.timeout(IO_TIMEOUT);
-    } catch (e) {
-      throw new Error(`Could not push the file within the given timeout ${IO_TIMEOUT}ms. ` +
-        `Original error: ${e.message}`);
-    }
+    await pushFile(service, relativePath, base64Data);
+  } catch (e) {
+    log.debug(e.stack);
+    throw new Error(`Could not push the file to '${remotePath}'.  Original error: ${e.message}`);
   } finally {
     service.close();
   }
+}
+
+async function deleteFileOrFolder (device, remotePath, isSimulator) {
+  return isSimulator
+    ? await deleteFromSimulator(device, remotePath)
+    : await deleteFromRealDevice(device, remotePath);
 }
 
 /**
@@ -353,16 +278,9 @@ async function pullFromRealDevice (device, remotePath, isFile) {
       throw new Error(`The requested path is not a folder. Path: '${remotePath}'`);
     }
 
-    if (fileInfo.isFile()) {
-      return await pullFileFromRealDevice(service, relativePath);
-    } else {
-      return await pullFolderFromRealDevice(service, relativePath);
-    }
-  } catch (e) {
-    if (e.message.includes(OBJECT_NOT_FOUND_ERROR_MESSAGE)) {
-      throw new Error(`Path '${remotePath}' does not exist on the device`);
-    }
-    throw e;
+    return fileInfo.isFile()
+      ? (await pullFile(service, relativePath)).toString('base64')
+      : (await pullFolder(service, relativePath)).toString();
   } finally {
     service.close();
   }
@@ -435,33 +353,6 @@ async function deleteFromRealDevice (device, remotePath) {
   }
 }
 
-/**
- * Get bundleIds which can mount by `--documents` flag
- *
- *
- * @param {Object} udid - The udid of the target device
- * @returns {Array<string>} A list of User level apps' bundle ids which has
- *                          'UIFileSharingEnabled' attribute.
- *                          Only user apps might have it.
- */
-async function getAvailableBundleIds (udid) {
-  const service = await services.startInstallationProxyService(udid);
-  try {
-    const applications = await service.listApplications({applicationType: 'User'});
-    const bundleIds = [];
-    for (const [key, value] of Object.entries(applications)) {
-      if (!value.UIFileSharingEnabled) {
-        continue;
-      }
-      bundleIds.push(key);
-    }
-    return bundleIds;
-  } finally {
-    service.close();
-  }
-}
-
-
 commands.pushFile = async function pushFile (remotePath, base64Data) {
   if (remotePath.endsWith('/')) {
     log.errorAndThrow(`It is expected that remote path points to a file and not to a folder. ` +
@@ -487,12 +378,6 @@ commands.pullFile = async function pullFile (remotePath) {
     : await pullFromRealDevice(this.opts.device, remotePath, true);
 };
 
-async function deleteFileOrFolder (device, remotePath, isSimulator) {
-  return isSimulator
-    ? await deleteFromSimulator(device, remotePath)
-    : await deleteFromRealDevice(device, remotePath);
-}
-
 commands.mobileDeleteFolder = async function mobileDeleteFolder (opts = {}) {
   let {remotePath} = opts;
   if (!remotePath.endsWith('/')) {
@@ -508,49 +393,6 @@ commands.mobileDeleteFile = async function mobileDeleteFile (opts = {}) {
                       `'${remotePath}' is given instead`);
   }
   return await deleteFileOrFolder(this.opts.device, remotePath, this.isSimulator());
-};
-
-commands.getSimFileFullPath = async function getSimFileFullPath (remotePath) {
-  let basePath = this.opts.device.getDir();
-  let appName = null;
-
-  if (this.opts.app) {
-    let appNameRegex = new RegExp(`\\${path.sep}([\\w-]+\\.app)`);
-    let appNameMatches = appNameRegex.exec(this.opts.app);
-    if (appNameMatches) {
-      appName = appNameMatches[1];
-    }
-  }
-  // de-absolutize the path
-  if (system.isWindows()) {
-    if (remotePath.indexof('://') === 1) {
-      remotePath = remotePath.slice(4);
-    }
-  } else {
-    if (remotePath.indexOf('/') === 0) {
-      remotePath = remotePath.slice(1);
-    }
-  }
-
-  if (remotePath.startsWith(appName)) {
-    let findPath = basePath;
-    if (!this.opts.platformVersion || util.compareVersions(this.opts.platformVersion, '>=', '8.0')) {
-      // the .app file appears in /Containers/Data and /Containers/Bundle both. We only want /Bundle
-      findPath = path.resolve(basePath, 'Containers', 'Bundle');
-    }
-    findPath = findPath.replace(/\s/g, '\\ ');
-
-    let { stdout } = await exec('find', [findPath, '-name', appName]);
-    let appRoot = stdout.replace(/\n$/, '');
-    let subPath = remotePath.substring(appName.length + 1);
-    let fullPath = path.resolve(appRoot, subPath);
-    log.debug(`Finding app-relative file: '${fullPath}'`);
-    return fullPath;
-  }
-
-  let fullPath = path.resolve(basePath, remotePath);
-  log.debug(`Finding sim-relative file: ${fullPath}`);
-  return fullPath;
 };
 
 commands.pullFolder = async function pullFolder (remotePath) {
