+++ lib/ios-deploy.js
@@ -1,4 +1,4 @@
-import { fs, timing, util } from 'appium-support';
+import { fs, timing } from 'appium-support';
 import path from 'path';
 import { services, utilities } from 'appium-ios-device';
 import B from 'bluebird';
@@ -6,10 +6,10 @@ import log from './logger';
 import _ from 'lodash';
 import { exec } from 'teen_process';
 import { extractBundleId } from './app-utils';
+import { pushFolder } from './ios-fs-helpers';
 
 const APPLICATION_INSTALLED_NOTIFICATION = 'com.apple.mobile.application_installed';
 const INSTALLATION_STAGING_DIR = 'PublicStaging';
-const DEFAULT_ITEM_PUSH_TIMEOUT_MS = 2 * 60 * 1000;
 const APPLICATION_NOTIFICATION_TIMEOUT_MS = 30 * 1000;
 const IOS_DEPLOY_TIMEOUT_MS = 4 * 60 * 1000;
 const IOS_DEPLOY = 'ios-deploy';
@@ -18,9 +18,6 @@ const APP_INSTALL_STRATEGY = Object.freeze({
   PARALLEL: 'parallel',
   IOS_DEPLOY,
 });
-// Mobile devices use NAND memory modules for the storage,
-// and the parallelism there is not as performant as on regular SSDs
-const MAX_PUSH_CHUNK_SIZE = 8;
 
 
 class IOSDeploy {
@@ -70,15 +67,20 @@ class IOSDeploy {
     if (_.toLower(strategy) === APP_INSTALL_STRATEGY.IOS_DEPLOY) {
       await installWithIosDeploy();
     } else {
+      const afcService = await services.startAfcService(this.udid);
       try {
-        const bundlePathOnPhone = await this.pushAppBundle(app, {
+        const bundleId = await extractBundleId(app);
+        const bundlePathOnPhone = path.join(INSTALLATION_STAGING_DIR, bundleId);
+        await pushFolder(afcService, app, bundlePathOnPhone, {
           timeoutMs: timeout,
           enableParallelPush: _.toLower(strategy) === APP_INSTALL_STRATEGY.PARALLEL,
         });
-        const bundleId = await extractBundleId(app);
         await this.installOrUpgradeApplication(bundlePathOnPhone, await this.isAppInstalled(bundleId));
       } catch (err) {
         log.warn(`Error installing app '${app}': ${err.message}`);
+        if (err instanceof B.TimeoutError) {
+          log.warn(`Consider increasing the value of 'appPushTimeout' capability`);
+        }
         log.warn(`Falling back to '${IOS_DEPLOY}' usage`);
         try {
           await installWithIosDeploy();
@@ -87,6 +89,8 @@ class IOSDeploy {
             `  - ${err.message}\n` +
             `  - ${err1.message}`);
         }
+      } finally {
+        afcService.close();
       }
     }
     log.info(`App installation succeeded after ${timer.getDuration().asMilliSeconds.toFixed(0)}ms`);
@@ -120,118 +124,6 @@ class IOSDeploy {
       installationService.close();
       notificationService.close();
     }
-  }
-
-  async pushAppBundle (app, opts = {}) {
-    const {
-      timeoutMs = DEFAULT_ITEM_PUSH_TIMEOUT_MS,
-      // We are pushing serially by default due to the issue https://github.com/appium/appium/issues/13115
-      enableParallelPush = false,
-    } = opts;
-
-    const timer = new timing.Timer().start();
-    const itemsToPush = await fs.glob('**', {
-      cwd: app,
-      nosort: true,
-      mark: true,
-    });
-    log.debug(`Successfully scanned the tree structure of '${path.basename(app)}' app`);
-    const [foldersToPush, filesToPush] = itemsToPush.reduce((acc, x) => {
-      acc[_.endsWith(x, path.sep) ? 0 : 1].push(x);
-      return acc;
-    }, [[], []]);
-    log.debug(`Got ${util.pluralize('folder', foldersToPush.length, true)} and ` +
-      `${util.pluralize('file', filesToPush.length, true)} to push`);
-    const afcService = await services.startAfcService(this.udid);
-    try {
-      const bundlePathOnPhone = await this.createAppPath(afcService, app);
-      // top-level folders must go first
-      const foldersToPushByHierarchy = foldersToPush
-        .sort((a, b) => a.split(path.sep).length - b.split(path.sep).length);
-      // create the folder structure first
-      for (const relativeFolderPath of foldersToPushByHierarchy) {
-        // createDirectory does not accept folder names ending with a path separator
-        const absoluteFolderPath = _.trimEnd(
-          path.join(bundlePathOnPhone, relativeFolderPath), path.sep
-        );
-        if (absoluteFolderPath) {
-          await afcService.createDirectory(absoluteFolderPath);
-        }
-      }
-      // do not forget about the root app folder
-      log.debug(`Successfully created the remote folder structure ` +
-        `(${util.pluralize('item', foldersToPush.length + 1, true)})`);
-
-      const pushFile = async (relativePath) => {
-        const absoluteSourcePath = path.join(app, relativePath);
-        const readStream = fs.createReadStream(absoluteSourcePath, {autoClose: true});
-        const absoluteDestinationPath = path.join(bundlePathOnPhone, relativePath);
-        const writeStream = await afcService.createWriteStream(absoluteDestinationPath, {
-          autoDestroy: true
-        });
-        writeStream.on('finish', writeStream.destroy);
-        let pushError = null;
-        const filePushPromise = new B((resolve, reject) => {
-          writeStream.on('close', () => {
-            if (pushError) {
-              reject(pushError);
-            } else {
-              resolve();
-            }
-          });
-          const onStreamError = (e) => {
-            readStream.unpipe(writeStream);
-            log.debug(e);
-            pushError = e;
-          };
-          writeStream.on('error', onStreamError);
-          readStream.on('error', onStreamError);
-        });
-        readStream.pipe(writeStream);
-        await filePushPromise.timeout(timeoutMs,
-          `Could not push '${absoluteSourcePath}' within the timeout of ${timeoutMs}ms. ` +
-          `Consider increasing the value of 'appPushTimeout' capability.`);
-      };
-
-      if (enableParallelPush) {
-        log.debug(`Proceeding to parallel files push (max ${MAX_PUSH_CHUNK_SIZE} writers)`);
-        const pushPromises = [];
-        for (const relativeFilePath of _.shuffle(filesToPush)) {
-          pushPromises.push(B.resolve(pushFile(relativeFilePath)));
-          // keep the push queue filled
-          if (pushPromises.length >= MAX_PUSH_CHUNK_SIZE) {
-            await B.any(pushPromises);
-          }
-          _.remove(pushPromises, (p) => p.isFulfilled());
-        }
-        if (!_.isEmpty(pushPromises)) {
-          // handle the rest of push promises
-          await B.all(pushPromises);
-        }
-      } else {
-        log.debug(`Proceeding to serial files push`);
-        for (const relativeFilePath of filesToPush) {
-          await pushFile(relativeFilePath);
-        }
-      }
-
-      log.debug(`Successfully pushed ${util.pluralize('app folder', foldersToPush.length, true)} ` +
-        `and ${util.pluralize('app file', filesToPush.length, true)} ` +
-        `within ${timer.getDuration().asMilliSeconds.toFixed(0)}ms`);
-      return bundlePathOnPhone;
-    } finally {
-      afcService.close();
-    }
-  }
-
-  async createAppPath (afcService, localAppPath) {
-    const basename = path.basename(localAppPath);
-    const relativePath = path.join(INSTALLATION_STAGING_DIR, basename);
-    try {
-      await afcService.deleteDirectory(relativePath);
-    } catch (ign) {}
-    await afcService.createDirectory(relativePath);
-    return relativePath;
   }
 
   async installApp (app, timeout) {
